+++ jobbergate-cli/CHANGELOG.md
@@ -3,7 +3,7 @@
 This file keeps track of all notable changes to jobbergate-cli
 
 ## Unreleased
-
+- Added clone command to job submissions so they can be resubmitted to the cluster when needed [PENG-1677, ASP-4597]
 
 ## 5.4.0a2 -- 2024-11-06
 ## 5.4.0a1 -- 2024-11-05